# DogAdoption/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An API with centralized dog adoption information.",
        "title": "DogAdoptionApp REST HATEOAS API",
        "contact": {
            "name": "Gustav Näslund, gn222gq",
            "email": "gn222gq@student.lnu.se"
        },
        "version": "1.0"
    },
    "host": "https://cscloud7-113.lnu.se/dogadoption",
    "basePath": "/api/v1",
    "paths": {
...
        "/dogs": {
            "get": {
                "description": "Retrieves a list of dogs based on provided query parameters like breed, size, and age.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dogs"
                ],
                "summary": "Get dogs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by dog breed",
                        "name": "breed",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by dog gender",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by if dog is neutered",
                        "name": "is_neutered",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by if dog is adopted",
                        "name": "is_adopted",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter dogs that are from a specific dog shelter",
                        "name": "shelter_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success, returns a list of dogs along with pagination details",
                        "schema": {
                            "$ref": "#/definitions/dogdto.DogsAndPaginationLinksDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request if the query parameters are invalid",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error if an error occurs while processing the request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds a new dog to the system with the provided dog data in JSON format. shelter_id field is for admins only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dogs"
                ],
                "summary": "Add a new dog",
                "parameters": [
                    {
                        "description": "Dog Data",
                        "name": "dog",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dogdto.NewDogDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success, returns the newly created dog information",
                        "schema": {
                            "$ref": "#/definitions/dogdto.DogDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request, if the JSON body cannot be parsed or mandatory fields are missing",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized, if the user does not have permission to add a dog",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error, if an error occurs while processing the request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/dogs/{id}": {
            "get": {
                "description": "Retrieves detailed information about a dog specified by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dogs"
                ],
                "summary": "Get a dog by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dog ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success, returns detailed information about the dog",
                        "schema": {
                            "$ref": "#/definitions/dogdto.DogDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found, if no dog matches the provided ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error, if an error occurs while processing the request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the information for an existing dog specified by its ID with the provided dog data in JSON format.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dogs"
                ],
                "summary": "Update dog information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dog ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dog Update Data",
                        "name": "dog",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dogdto.UpdateDogDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success, returns the updated dog information",
                        "schema": {
                            "$ref": "#/definitions/dogdto.DogDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request, if the JSON body cannot be parsed, mandatory fields are missing, or the dog data is incomplete",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized, if the user does not have permission to update the dog",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error, if an error occurs while processing the request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a dog specified by its ID if the requester has the necessary permissions.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dogs"
                ],
                "summary": "Delete a dog",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dog ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Dog deleted successfully"
                    },
                    "400": {
                        "description": "Bad Request if the request was malformed",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized if the user does not have permission to delete the dog",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found if the dog with the specified ID does not exist",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error for any server errors",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dogdto.DogDTO": {
            "type": "object",
            "properties": {
                "adoption_fee": {
                    "type": "integer"
                },
                "birth_date": {
                    "type": "string"
                },
                "breed": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "friendly_with": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "is_adopted": {
                    "type": "boolean"
                },
                "is_neutered": {
                    "type": "boolean"
                },
                "links": {
                    "$ref": "#/definitions/dogdto.DogLinksDTO"
                },
                "name": {
                    "type": "string"
                },
                "shelter_id": {
                    "type": "integer"
                }
            }
        },
        "dogdto.DogLinksDTO": {
            "type": "object",
            "properties": {
                "self_link": {
                    "type": "string"
                },
                "shelter_link": {
                    "type": "string"
                }
            }
        },
        "dogdto.DogsAndPaginationLinksDTO": {
            "type": "object",
            "properties": {
                "dogs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dogdto.DogDTO"
                    }
                },
                "pagination_links": {
                    "$ref": "#/definitions/dto.PaginationLinksDTO"
                }
            }
        },
        "dogdto.NewDogDTO": {
            "type": "object",
            "properties": {
                "adoption_fee": {
                    "type": "integer"
                },
                "birth_date": {
                    "type": "string"
                },
                "breed": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "friendly_with": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "is_adopted": {
                    "type": "boolean"
                },
                "is_neutered": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "shelter_id": {
                    "type": "integer"
                }
            }
        },
        "dogdto.UpdateDogDTO": {
            "type": "object",
            "properties": {
                "adoption_fee": {
                    "type": "integer"
                },
                "birth_date": {
                    "type": "string"
                },
                "breed": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "friendly_with": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "is_adopted": {
                    "type": "boolean"
                },
                "is_neutered": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "dto.PaginationLinksDTO": {
            "type": "object",
            "properties": {
                "first": {
                    "type": "string"
                },
                "last": {
                    "type": "string"
                },
                "next": {
                    "type": "string"
                },
                "previous": {
                    "type": "string"
                },
                "self": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}